requirements.txt
streamlit
PyPDF2
langchain
langchain-core
langchain-huggingface
python-docx
fpdf
python-dotenv
